refactor: move SetupMppPaymentViewerUseCase into commonMain
