--- wallet-sdk-ui/src/iosMain/kotlin/com/michaeltchuang/walletsdk/ui/liquidStream/domain/usecases/SetupMppPaymentViewerUseCase.ios.kt
@@ -1,74 +1,0 @@
-package com.michaeltchuang.walletsdk.ui.liquidStream.domain.usecases
-
-import com.michaeltchuang.walletsdk.core.network.model.AlgorandNetwork
-import com.michaeltchuang.walletsdk.core.railmpp.MppNetworks
-import com.michaeltchuang.walletsdk.core.railmpp.domain.model.ChatMessage
-import com.michaeltchuang.walletsdk.core.railmpp.domain.model.ConsentApproval
-import com.michaeltchuang.walletsdk.core.railmpp.domain.model.ConsentTerms
-import com.michaeltchuang.walletsdk.core.railmpp.domain.repository.MppWalletSigner
-import com.michaeltchuang.walletsdk.core.railmpp.domain.usecase.GetSessionVaultConfigUseCase
-import com.michaeltchuang.walletsdk.core.railmpp.smartcontract.EscrowSessionVaultManagerClient
-import com.michaeltchuang.walletsdk.ui.liquidStream.domain.manager.MppPaymentViewerManager
-import com.michaeltchuang.walletsdk.ui.liquidStream.domain.transport.CallbackRtcDataChannel
-import io.github.aakira.napier.Napier
-import kotlinx.coroutines.CoroutineScope
-
-actual class SetupMppPaymentViewerUseCase actual constructor(
-    private val viewerManager: MppPaymentViewerManager,
-    private val getSessionVaultConfigUseCase: GetSessionVaultConfigUseCase,
-) {
-    companion object {
-        private const val TAG = "SetupMppPaymentViewer"
-    }
-
-    data class Params(
-        val dataChannel: CallbackRtcDataChannel,
-        val viewerAddress: String,
-        val hostAddress: String,
-        val scope: CoroutineScope,
-        val signer: MppWalletSigner,
-        val mppNetwork: String,
-        val requestMppConsent: suspend (ConsentTerms) -> ConsentApproval,
-        val setViewerSessionVaultProgress: (remainingBalanceMicroUsdc: Long, progressBalanceMicroUsdc: Long) -> Unit,
-        val signFido2Challenge: suspend (challenge: ByteArray, address: String) -> ByteArray?,
-        val onChatMessageReceived: (ChatMessage) -> Unit = {},
-    )
-
-    operator fun invoke(params: Params) {
-        val viewerAddress =
-            params.viewerAddress.takeIf { it.isNotBlank() }
-                ?: run {
-                    Napier.w("[VIEWER_MPP_SETUP_SKIP] reason=blank_viewer host=${params.hostAddress}", tag = TAG)
-                    return
-                }
-        val sessionVaultNetwork = params.mppNetwork.toAlgorandNetwork()
-        EscrowSessionVaultManagerClient.configureForNetwork(sessionVaultNetwork)
-        val sessionVaultConfig = getSessionVaultConfigUseCase(sessionVaultNetwork)
-        viewerManager.start(
-            MppPaymentViewerManager.StartParams(
-                dataChannel = params.dataChannel,
-                viewerAddress = viewerAddress,
-                scope = params.scope,
-                signer = params.signer,
-                mppNetwork = params.mppNetwork,
-                sessionVaultAppId = sessionVaultConfig.appId,
-                requestMppConsent = params.requestMppConsent,
-                setViewerSessionVaultProgress = params.setViewerSessionVaultProgress,
-                signFido2Challenge = params.signFido2Challenge,
-                onChatMessageReceived = params.onChatMessageReceived,
-            ),
-        )
-    }
-
-    private fun String.toAlgorandNetwork(): AlgorandNetwork =
-        when {
-            this == MppNetworks.ALGORAND_MAINNET || contains("mainnet", ignoreCase = true) -> AlgorandNetwork.MAINNET
-            this == MppNetworks.ALGORAND_FUTURENET ||
-                contains(
-                    "futurenet",
-                    ignoreCase = true,
-                ) ||
-                contains("fnet", ignoreCase = true) -> AlgorandNetwork.FUTURENET
-            else -> AlgorandNetwork.TESTNET
-        }
-}
